requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
alpha_vantage==3.0.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
huggingface-hub==0.26.3
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.9.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.5
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
tokenizers==0.20.3
torch>=2.5.0
tqdm==4.67.1
transformers==4.46.3
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.18.3
nvidia-cudnn-cu12>=9.0.0,<9.2.0
nvidia-cublas-cu12==12.4.5.8